create variables to colors
