+++ react-native/training/src/app/screens/TodoList/components/Checkbox/styles.js
@@ -1,15 +1,17 @@
 import { StyleSheet } from 'react-native';
 
+import { black, blackTransparent } from '../../../../../constants/colors';
+
 export default StyleSheet.create({
   box: {
     height: 20,
     width: 20,
     borderWidth: 2,
-    borderColor: 'black'
+    borderColor: black
   },
   inner: {
     flex: 1,
     margin: 2,
-    backgroundColor: 'rgba(0,0,0,0.8)'
+    backgroundColor: blackTransparent
   }
 });
