+++ react-native/training/src/app/screens/TodoList/components/Footer/styles.js
@@ -1,11 +1,13 @@
 import { StyleSheet } from 'react-native';
 
+import { chestnutRose } from '../../../../../constants/colors';
+
 export default StyleSheet.create({
   footer: {
     paddingVertical: 15,
     alignItems: 'center'
   },
   remove: {
-    color: '#CD5C5C'
+    color: chestnutRose
   }
 });
