+++ react-native/training/src/app/screens/TodoList/components/Title/styles.js
@@ -1,12 +1,14 @@
 import { StyleSheet } from 'react-native';
 
+import { skyblue, white } from '../../../../../constants/colors';
+
 export default StyleSheet.create({
   header: {
-    backgroundColor: 'skyblue',
+    backgroundColor: skyblue,
     padding: 15
   },
   title: {
     textAlign: 'center',
-    color: 'white'
+    color: white
   }
 });
