+++ react-native/training/src/app/screens/TodoList/styles.js
@@ -1,11 +1,13 @@
 import { StyleSheet } from 'react-native';
 
+import { whitesmoke } from '../../../constants/colors';
+
 export default StyleSheet.create({
   container: {
     flex: 1
   },
   divider: {
     height: 1,
-    backgroundColor: 'whitesmoke'
+    backgroundColor: whitesmoke
   }
 });
