+++ react-native/training/src/constants/colors.js
@@ -1,8 +1,14 @@
 exports.blue = '#288CCB';
 exports.green = '#53E69D';
+exports.chestnutRose = '#CD5C5C';
 
 exports.transparent = 'transparent';
+exports.whitesmoke = 'whitesmoke';
+exports.skyblue = 'skyblue';
+exports.white = 'white';
+exports.black = 'black';
 
 exports.white = '#FFF';
 exports.black = '#000';
 exports.gray = 'rgba(0, 0, 0, 0.38)';
+exports.blackTransparent = 'rgba(0,0,0,0.8)';
