+++ react-native/training/src/redux/todoList/reducer.js
@@ -1,40 +1,11 @@
 import { createReducer } from 'redux-recompose';
 import Immutable from 'seamless-immutable';
 
+import reducerEffects from './effects';
 import { actionsTypes } from './actions';
 
 const initialState = {
   items: []
-};
-
-const reducerEffects = {
-  addItem: (state, action) => ({
-    ...state,
-    [action.target]: [
-      {
-        label: action.payload,
-        completed: false
-      },
-      ...state.items
-    ]
-  }),
-  removeItem: (state, action) => ({
-    ...state,
-    [action.target]: state.items.filter((item, i) => i !== action.payload)
-  }),
-  toggleItemCompleted: (state, action) => ({
-    ...state,
-    [action.target]: state.items.map((item, i) => {
-      if (i === action.payload) {
-        return { label: item.label, completed: !item.completed };
-      }
-      return item;
-    })
-  }),
-  removeCompleted: (state, action) => ({
-    ...state,
-    [action.target]: state.items.filter(item => !item.completed)
-  })
 };
 
 const reducerDescription = {
